refactor: move classes to own folder
