+++ client/classes/Bitmap.js
