+++ client/classes/Square.js
